Added a check that API docs have changed before trying to commit

+++ src/main/scala/release/tasks.scala
@@ -379,33 +379,39 @@ case object tasks {
     val destBase   = ghpagesDir / "docs" / "api"
     val destVer    = destBase / version.value
 
-    log.info(s"Copying ${generatedDocs} to ${destVer}")
+    log.info(s"Copying ${generatedDocs.relPath(git.workingDir)} to <gh-pages>/${destVer.relPath(ghpagesDir)}")
     if (destVer.exists) IO.delete(destVer)
     IO.copyDirectory(generatedDocs, destVer, overwrite = true)
 
-    ghpagesGit.stageAndCommit(s"API docs v${git.version}")(destVer)
-    log.info(s"Committed ${destVer}")
+    if (! ghpagesGit.isDirty) {
+      // If there are no changes we don't do anything else
+      log.warn("No changes to commit and publish")
 
-    if (latest) {
-      // NOTE: .nojekyll file is needed for symlinks (see https://github.com/isaacs/github/issues/553)
-      val _nojekyll = ghpagesDir / ".nojekyll"
-      if (! _nojekyll.exists) {
-        log.info(s"Adding .nojekyll file")
-        IO.write(_nojekyll, "")
+    } else {
+      ghpagesGit.stageAndCommit(s"API docs v${git.version}")(destVer)
+      log.debug(s"Committed ${destVer}")
 
-        ghpagesGit.stageAndCommit("Added .nojekyll file for symlinks")(_nojekyll)
+      if (latest) {
+        // NOTE: .nojekyll file is needed for symlinks (see https://github.com/isaacs/github/issues/553)
+        val _nojekyll = ghpagesDir / ".nojekyll"
+        if (! _nojekyll.exists) {
+          log.info(s"Adding .nojekyll file")
+          IO.write(_nojekyll, "")
+
+          ghpagesGit.stageAndCommit("Added .nojekyll file for symlinks")(_nojekyll)
+        }
+
+        val destLatest = destBase / "latest"
+        log.info(s"Symlinking ${destLatest.relPath(ghpagesDir)} to ${destVer.relPath(ghpagesDir)}")
+        Files.deleteIfExists(destLatest.absPath)
+        Files.createSymbolicLink(destLatest.absPath, destVer.relPath(destLatest.getParentFile))
+
+        ghpagesGit.stageAndCommit(s"Symlinked ${git.version} as latest")(destLatest)
       }
 
-      val destLatest = destBase / "latest"
-      log.info(s"Symlinking ${destLatest} to ${destVer}")
-      Files.deleteIfExists(destLatest.absPath)
-      Files.createSymbolicLink(destLatest.absPath, destVer.relPath(destLatest.getParentFile))
-
-      ghpagesGit.stageAndCommit(s"Symlinked ${git.version} as latest")(destLatest)
+      log.info("Publishing API docs...")
+      ghpagesGit.push(url)(gh_pages).critical
     }
-
-    log.info("Publishing API docs...")
-    ghpagesGit.push(url)(gh_pages).critical
   }
 
 }
